#59 - Made InMemoryStorage thread-safe and moved list() methods tests from InMemoryStorageTest to InMemoryStorageListTest
