+++ src/test/java/com/artipie/asto/InMemoryStorageListTest.java
@@ -1,0 +1,74 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2020 artipie.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package com.artipie.asto;
+
+import com.artipie.asto.blocking.BlockingStorage;
+import com.artipie.asto.memory.InMemoryStorage;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.stream.Collectors;
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.Matchers;
+import org.junit.jupiter.api.Test;
+
+/**
+ * Tests for {@link InMemoryStorage#list(Key)}.
+ *
+ * @since 0.14
+ */
+public final class InMemoryStorageListTest {
+
+    @Test
+    void shouldListNoKeysWhenEmpty() {
+        final BlockingStorage blocking = new BlockingStorage(this.storage());
+        final Collection<String> keys = blocking.list(new Key.From("a", "b"))
+            .stream()
+            .map(Key::string)
+            .collect(Collectors.toList());
+        MatcherAssert.assertThat(keys, Matchers.empty());
+    }
+
+    @Test
+    void shouldListKeysInOrder() {
+        final byte[] data = "some data!".getBytes();
+        final BlockingStorage blocking = new BlockingStorage(this.storage());
+        blocking.save(new Key.From("1"), data);
+        blocking.save(new Key.From("a", "b", "c", "1"), data);
+        blocking.save(new Key.From("a", "b", "2"), data);
+        blocking.save(new Key.From("a", "z"), data);
+        blocking.save(new Key.From("z"), data);
+        final Collection<String> keys = blocking.list(new Key.From("a", "b"))
+            .stream()
+            .map(Key::string)
+            .collect(Collectors.toList());
+        MatcherAssert.assertThat(
+            keys,
+            Matchers.equalTo(Arrays.asList("a/b/2", "a/b/c/1"))
+        );
+    }
+
+    private InMemoryStorage storage() {
+        return new InMemoryStorage();
+    }
+}
